Discord-Bot/requirements.txt
discord.py==2.3.1
PyNaCl==1.5.0
yt-dlp==2023.7.6
python-dotenv==1.0.0
